requirements.txt
pandas==2.2.2
numpy==1.26.4
psycopg[binary]==3.2.1
python-dotenv==1.0.1
tqdm==4.66.4
